Update set-up-dependencies.sh
easyocr wasn't installed after deploying 6.6.1 (unsure if this affects earlier versions).

Changing python venv version seems to resolve this since python version installed is 3.13 not 3.11. Running set-up-dependencies.sh in the container generated a bunch of errors that start w/ the wrong version of python3-venv, updating it resolved the errors and allowed it to build properly.